+++ make.ps1
@@ -22,30 +22,74 @@
     $ini["DEFAULT"]
 }
 
-$config = Parse-IniFile(".\config.ini")
-$config["project_path_linux"] = "/mnt/$($config["project_path"].Substring(0,1).ToLower())/$($config["project_path"].Substring(3))"
+function Normalize-Path($path) {
+    return $path -replace '\\+', '/'
+}
 
+function _To-Linux-Path($path) {
+    $matches = @{}
+    if ($path -match '^([A-z])\:(.*)$') {
+        return "/mnt/$($matches[1].ToLower())$(Normalize-Path $matches[2])"
+    }
+    $URL_Format_Error = New-Object System.FormatException "Wrong Windows path format: $($path)"
+    Throw $URL_Format_Error
+}
 
-function ESP32-Make-Command($command) {
+function _To-Windows-Path($path) {
+    $matches = @{}
+    if ($path -match '^/mnt/([A-z])(.*)$') {
+        return "$($matches[1]):$($matches[2])"
+    }
+    $FormatError = New-Object System.FormatException "Wrong Linux path format (must start with /mnt/): $($path)"
+    Throw $FormatError
+}
+
+function _Read-Config {
+    $config = Parse-IniFile(".\config.ini")
+    $config["project_path"] = Normalize-Path $config["project_path"]
+    $config["idf_path"] = Normalize-Path $config["idf_path"]
+    $config["idf_path_linux"] = _To-Linux-Path $config["idf_path"]
+    $config["project_path_linux"] = _To-Linux-Path $config["project_path"]
+    $config["script_path"] = _To-Linux-Path $PSScriptRoot
+    $config["output_file"] = "$($config["script_path"])/output.log"
+
+    return $config
+}
+
+function _Get-Last-Output {
+    Get-content -tail 1 (_To-Windows-Path $config["output_file"])
+}
+
+$config = _Read-Config
+
+function _Make-Flash-Command {
+    $cmd = _Get-Last-Output
+    if (!$cmd.StartsWith("python")) {
+        $Flash_Error = New-Object System.ApplicationException "Flasher command not found. Did the make command run successfull?"
+        Throw $Flash_Error
+    }
+    $cmd = $cmd -replace $config["idf_path_linux"], $config["idf_path"]
+    $cmd = $cmd -replace $config["project_path_linux"], $config["project_path"]
+
+    return $cmd
+}
+
+function _ESP32-Make-Command($command) {
     if ($command -eq "all") {
         $command = "all -j$($config["threads"])"
     }
-    bash -l -c "cd $($config["project_path_linux"]) && make $($command)"
+    $cmd = "bash -l -c `"cd $($config["project_path_linux"]) && make $($command) | tee $($config["output_file"])`""
+    Invoke-Expression $cmd
  }
 
-function ESP32-Flash-Command {
-    ESP32-Make-Command "all"
-    $cmd =  "python $($config["idf_path"])/components/esptool_py/esptool/esptool.py " +
-        "--chip esp32 --port $($config["port"]) --baud $($config["flash_baud"]) --before esp32r0 " +
-        "--after hard_reset write_flash -u --flash_mode dio --flash_freq 40m --flash_size detect " +
-        "0x1000 $($config["project_path"])/build/bootloader/bootloader.bin " +
-        "0x10000 $($config["project_path"])/build/p1p.bin " +
-        "0x8000 $($config["project_path"])/build/partitions_singleapp.bin"
+function _ESP32-Flash-Command {
+    _ESP32-Make-Command "all"
+    $cmd = _Make-Flash-Command
 
-    Invoke-Expression $cmd
+    Invoke-Expression $cmd    
 }
 
-function ESP32-Monitor-Command {
+function _ESP32-Monitor-Command {
     cmd.exe /c "python -m serial.tools.miniterm --rts 0 --dtr 0 --raw $($config["port"]) $($config["monitor_baud"])"
 }
 
@@ -57,9 +101,9 @@ function ESP32-Make() {
 
     foreach($arg in $args) {
         if ($arg -eq "flash") {
-            $cmd = "$($cmd); ESP32-Flash-Command"
+            $cmd = "$($cmd); _ESP32-Flash-Command"
         } elseif ($arg -eq "monitor") {
-            $cmd = "$($cmd); ESP32-Monitor-Command"
+            $cmd = "$($cmd); _ESP32-Monitor-Command"
         } elseif ($arg -eq "help") {
             write-host ""
             write-host "Welcome to Espressif IDF build system porting to PowerShell. Some useful make targets:" -foreground "green"
@@ -82,7 +126,7 @@ function ESP32-Make() {
 	        write-host "'make partition_table', etc, etc." -foreground "green"
 	        write-host ""
         } else {
-            $cmd = "$($cmd); ESP32-Make-Command $($arg)"
+            $cmd = "$($cmd); _ESP32-Make-Command $($arg)"
         }
     }